% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{browser_addin}
\alias{browser_addin}
\title{RStudio AddIn.}
\usage{
browser_addin(data)
}
\arguments{
\item{data}{Container with artifacts; see \code{\link[repository:read_artifacts]{repository::read_artifacts()}}
for more details.}
}
\description{
\code{browser_addin} runs the \code{htmlwidget} implemented in this
package as a RStudio AddIn (see \link[shiny:runGadget]{shiny::runGadget} for details).
}
