% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{prepare_plots}
\alias{prepare_plots}
\title{Internal widget API.}
\usage{
prepare_plots(data, embed = is_knitr())
}
\arguments{
\item{data}{container of artifacts; see \code{\link[repository:is_container]{repository::is_container()}}.}

\item{embed}{Whether to embed plots as base64-encoded strings.}
}
\value{
A \code{list} with two elements: the \code{data} container possibly with
embedded plots; widget dependencies \code{html_deps}.
}
\description{
\code{prepare_plots} processes the \code{steps} list and either:
\itemize{
\item extracts plots into \code{png} files which are then attached to the
html widget via \code{dependencies} (see \link[htmltools:htmlDependency]{htmltools::htmlDependency})
\item embeds the png bitmap as base64-encoded string under the \code{contents}
key
}

\code{png} files are stored in a temporary directory persisting throughout
the R session. Subsequent calls to \code{prepare_plots} will only process
new plots and thus be faster.
}
