% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{create_widget}
\alias{create_widget}
\title{Widget API.}
\usage{
create_widget(data)
}
\arguments{
\item{data}{container of artifacts; see \link[repository:is_container]{repository::is_container}.}
}
\value{
A \code{list} with two elements: the \code{data} container possibly with
embedded plots; widget dependencies \code{html_deps}.
}
\description{
\code{prepare_plots} pre-processes the list of artifacts and
either:
\itemize{
\item extracts plots into \code{png} files which are then attached to the
html widget via \code{dependencies} (see \link[htmltools:htmlDependency]{htmltools::htmlDependency})
\item embeds the png bitmap as base64-encoded string under the \code{contents}
key
}

\code{png} files are stored in a temporary directory persisting throughout
the R session. Subsequent calls to \code{prepare_plots} will only process
new plots and thus be faster.
}
